Selenium IDE: Running v4 without v3 (#1258)
* verify not editable tests disabled (#1151)

changed to : 'verify not editable' = has been disabled
rather than what the original said 
hasn't been disabled

* [v4] Attempting to run v3 inside of v4

* chromedriver install via electron-chromedriver

* move renderer out into a separate package

in addition, this adds electron-tabs to the frontend

* remove unused chromedriver install script

* making selenium ide v3 work with install stuff

it is slow as hell though, so there is that

* remove nodeIntegration, and instead lay boilerplate towards a channel system

I would feel really bad if someone used selenium-ide, and some zero day in electron / chrome let them leverage the full NodeJS api.
So the plan would be to swap that out for a channel system to leverage enhanced permissions.

* selenium ide v3 renders and some code cleanup

* first time rendering v3 in electron

the ipc layer is one directional and everything is kinda hacky right now
loading and saving side files is the only part that really works
next step is to make it so electron can talk into webview via ipc
and then get recording working like it used to

* working bi-directional api

and it feels pretty good to work with :D

* voraciously cleaning up this stuff

* progress on side extension polyfill
I am starting to handle the tabs.onUpdated event
It isnt a lot, but it sets the boilerplate

* v4 an almost insignificant amount of docs
event linked to corresponding chrome api

* extending type enforcement through the api
wah wah wee wah

* making types slightly more consistent and bug squashing

yay types!

* trying to move to webdriver stuff

* trying to get playback to use webdriver

* Breaking everything to be "cleaner"
Goodbye type completion, goodbye anything working.

This is kind of a bad idea, but the nice thing is that
it aligns the whole package around its true function at this point:
providing a polyfill for the webextension api that integrates with
electron. If we have that, we have v3 and can start making v4 changes.

* oh my sweet types, how I have missed you

* oh boy things work again

* [v4] save what i have

* Replacing a polyfill with an api

* selenium ide v4 but without the polyfill

* wiring in rpc style broadcastng to keep browser windows in sync, overengineering

* 0 linter errors, server and client have shared state now

* fully typing runtime, commons, and adding to model

* configurable driver and startup ux

* getting electron driver working again

* A basic spark, fun with panels

* style and building better

* PLAYBACK!!! :D :D :D

* re-running playback multiple times works

* latest code shape

* repairing build command

* clean watch and build commands

* consistentizing build and watch commands just a bit more

* passing types in side-migrate

* module file for unescape

* atomic and stable build and watch commands

* cleaning up last ts complaints

* Changed casing (#1330)

* Fix stale frontend (#1329)

* drop unneeded loggers for now
some of these should perhaps be console.debug, but w/e

* fix stale frontend
we were not memoizing on enough things which caused a prev
version of test and command to be cached, causing things
like test or command updates to not reflect

* thin down command list panel
less room there is more room for playback windows

* hide panels less aggressively
if a new playback window is opened, hiding panels
feels weird and somewhat unseemly. This hides panels
only if all windows are unfocused, which means
you are not using the app atm

* reduce build memory requirements

* webpack rapid build

* fix invalid capitalization

* playback progress and re-collapsing ux

* latest everything

* fixed recurring playback

* addl ux work

* add record placeholder function

* typing side recorder postprocessor

* recorder is mostly typed

* recording is so much closer

* commands are piping whoo!

* getting better every day :3

* updating mui, collapsing components

* tabs layer perhaps?

* suite editor system

* project editing tab

* recording concats steps now

* tuning some scrolling stuff

* save and load controls

* rebuild driver between projects

* typed argument fields

* resolve type errors

* renaming things, wip towards select control

* select element :D :D

* inspect elements option in all playback windows

Co-authored-by: OzBob <OzBob@users.noreply.github.com>
Co-authored-by: Harsh <hsengar89@gmail.com>